Update config_flow.py
I propose to add an optional field for filling out the App key if it is already created on the Octoprint instance. This may help people with network discovery issues. If the key is filled in, then the discovery function will be skipped. As I'm not knowledgeable in Python I don't know if the condition "if CONF_API_KEY in user_input:" will return false when the CONF_API_KEY is "". If not, then the variable should be initialised with None instead of "" but I can just suppose as I can't test it myself.