Update tags for the different forecast modes for the OpenWeatherMap API
Before, the free API call was apparently a daily forecast. Now the free call is for a 3-hour forecast for 5 days but documentation for this integration is not clear about it. All other calls are paid. This is confusing for someone who just subscribed to the API for starting using it with HA.

Text is changed to reflect the descriptions found here: https://openweathermap.org/price